Preparation for audio muxing (interleave).
Now isom_write_sample once pools the sample recieved for a while rather than write it immediately, and actually write them after fulfilled designated duration.
Currently, the pooling causes duplication of sample, which is not effective algorithm.
And much of trak system is not effective too.
These should be fixed in the future.
Note also that current interleve system does not suit for highly intermittent streams, such as 0.1 fps video.
To fix this, inter-trak-arbitration method is required for interleaving.