Set one on roll_distance when The first picture of the group is substantially an IDR picture.
The intra picture that doesn't lead all pictures that depend on former groups is substantially an instantaneous decoding picture.
Therefore, roll_distance of the group that includes this picture as the first sample is zero.
However, this means this group cannot have random access point, since roll_distance = 0 must not be used. This is not preferable.
So we treat this group as 'roll' and pick roll_distance = 1 though this treatment is over-estimation.