Add experimental UTF-8 BOM handling on chapter strings.
This feature adds BOM to the chapter strings in the chapter list.
It makes MKVToolNix recognize the UTF-8 charset by default even on Windows.
Apparently LAV Splitter, VLC media player, and MediaInfo works fine,
but there may be an application that dislikes the BOM.
Use boxdumper --chapter to extract chapter lists with BOM.