bstream: Remove redundant casts and AND operations.
According to C99 6.3.1.3 Signed and unsigned integers,
if the new type is unsigned, the value is converted by repeatedly adding
or subtracting one more than the maximum value that can be represented in
the new type until the value is in the range of the new type.

Therefore, for instance,
  uint8_t c = 0x280;
  printf( "%"PRIx8, c );
the above codes display 0x80 since any value greater than 0xFF with all
lower 8 bits equal to 0 is a multiple of (0xFF + 1) and you get 0x80 by
subtracting (0xFF + 1) from 0x280 twice.