box: Initialize predefined children by non-existing box constants whenever allocating container boxes.
Check box existence with the new methods in order to prevent you from
missing null deferencing checks. Now, you can do so somewhat safer coding
than the previous. Be careful not to overwrite constants which are read
only. It's a new issue because of this change, but I think it's still
safer on the whole.